+++ tests/services/test_config_handler.py
@@ -2,4 +2,3 @@ from os import path
 from dotpyle.services.config_handler import ConfigHandler
 import pytest
 from subprocess import CalledProcessError, SubprocessError
-from dotpyle_mock_config_files import dotpyle_ok_cases, dotpyle_error_cases
