+++ tests/unitary/services/test_config_checker.py
@@ -1,0 +1,21 @@
+from os import path
+import pytest
+from tests.utils.mocks.config_valid_cases import valid_cases
+from tests.utils.mocks.config_invalid_cases import invalid_cases
+from dotpyle.services.config_checker import ConfigChecker
+
+
+@pytest.mark.parametrize("config", [*valid_cases])
+def test_check_config_valid_configs(config):
+    checker = ConfigChecker()
+
+    errors = checker.check_config(config)
+    assert len(errors) == 0
+
+
+@pytest.mark.parametrize("config", [*invalid_cases])
+def test_check_config_invalid_configs(config):
+    checker = ConfigChecker()
+
+    errors = checker.check_config(config)
+    assert len(errors) > 0
