+++ tests/utils/mocks/config_invalid_cases.py
@@ -1,0 +1,8 @@
+invalid_cases = [
+    {
+        "version": -1,
+    },
+    {
+        "version": 0,
+    },
+]
